refactor source file organization
