+++ lib/ui/components/current_timer/CurrentTimerStopButton.svelte
