+++ lib/ui/components/reports/ReportBlockHeader.svelte
