+++ lib/ui/components/reports/TogglReportProjectList.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
-	import type Report from 'lib/model/Report';
+	import type { Report } from 'lib/model/Report';
 	import millisecondsToTimeString from 'lib/util/millisecondsToTimeString';
 	import { dailySummary } from 'lib/util/stores';
 	import { onDestroy } from 'svelte';
 
 	let list: { color: string; duration: string; name: string }[];
 
-	const computeList = (r: Report) => {
+	const computeList = (r: Report<any>) => {
 		if (r == null) {
 			return [];
 		}
