+++ lib/ui/components/reports/charts/BarChart.svelte
