+++ rollup.config.js
@@ -33,5 +33,9 @@ export default {
 			emitCss: false,
 			preprocess: autoPreprocess()
 		})
-	]
+	],
+	onwarn: function (warning, warn) {
+    if (warning.code === 'CIRCULAR_DEPENDENCY') return;
+    warn(warning);
+}
 };
